Separate task class from module.h
